Add validation for Product/Delete

+++ SportsStore/Controllers/ProductController.cs
@@ -103,6 +103,8 @@ namespace SportsStore.Controllers {
         public IActionResult DeleteConfirmed(int id) {
             try {
                 Product product = _productRepository.GetById(id);
+                if (product == null)
+                    return NotFound();
                 _productRepository.Delete(product);
                 _productRepository.SaveChanges();
                 TempData["message"] = $"You successfully deleted product {product.Name}.";
